+++ Arma3ObjectBuilder/ui/props_object_mesh.py
@@ -564,7 +564,7 @@ class A3OB_PT_object_mesh(bpy.types.Panel):
         
         if object_props.is_a3_lod:
             layout.prop(object_props, "lod", text="Type")
-            if int(object_props.lod) in data.lod_resolution_position.keys():
+            if int(object_props.lod) in data.lod_resolution:
                 layout.prop(object_props, "resolution")